Add zsys handling for booting on older partitions
- A special org.zsys:bootfs property enables us to know that a given
system has a zsys structure. Only datasets with org.zsys:bootfs under
<poolname>/ROOT/ will be mounted.
- Booting from a snapshot (rollbacking) creates a separate root dataset,
suffixed by a newly generated -<uid>. This clone will be started,
mountpoints will be switched from "noauto" to "on" (on datasets that don't
have org.zsys:bootfs) and previous "on" to "noauto".
Datasets with "off" aren't changed.
- If new system is booted, it becomes "current", and zsys will commit
it.
- Older version with -<uid> will be garbage collected at some point.
- We are using real mountpoints, so that zfs mount -a mounts everything
in altroot as needed.